validating data in form using validation middleware
